+++ qolmat/imputations/imputers_pytorch.py
@@ -1,35 +1,17 @@
-from typing import Dict, List, Optional, Tuple
-
-import math
-
-from sklearn.base import BaseEstimator
-from sklearn import preprocessing
-
+from typing import Dict, List, Optional
 import numpy as np
 import pandas as pd
-import time
-import gc
 
-import torch
-from torch.utils.data import DataLoader, TensorDataset
-
+from sklearn.base import BaseEstimator
 
 from qolmat.imputations.imputers import ImputerRegressor, ImputerGenerativeModel
 from qolmat.utils.exceptions import PytorchNotInstalled
-from qolmat.benchmark import missing_patterns, metrics
+from qolmat.benchmark import metrics
 
-# try:
-#     from tensorflow.keras.callbacks import EarlyStopping
-# except ModuleNotFoundError:
-#     raise PytorchNotInstalled
-
-
-def get_num_params(model):
-    model_parameters = filter(lambda p: p.requires_grad, model.parameters())
-    params = sum([np.prod(p.size()) for p in model_parameters])
-
-    return params
-
+try:
+    import torch
+except ModuleNotFoundError:
+    raise PytorchNotInstalled
 
 class ImputerRegressorPytorch(ImputerRegressor):
     def __init__(
@@ -58,17 +40,28 @@ class ImputerGenerativeModelPytorch(ImputerGenerativeModel):
         model: Optional[BaseEstimator] = None,
         epochs: int = 100,
         batch_size: int = 100,
+        x_valid: pd.DataFrame = None,
+        x_valid_mask: pd.DataFrame = None,
         print_valid: bool = False,
+        metrics_valid: Dict[str, callable] = {
+            "mae": metrics.mean_absolute_error
+        },
         **hyperparams,
     ):
         super().__init__(groups=groups, model=model, **hyperparams)
         self.epochs = epochs
         self.batch_size = batch_size
+        self.x_valid = x_valid
+        self.x_valid_mask = x_valid_mask
         self.print_valid = print_valid
+        self.metrics_valid = metrics_valid
 
     def get_params_fit(self) -> Dict:
         return {
             "epochs": self.epochs,
             "batch_size": self.batch_size,
+            "x_valid": self.x_valid,
+            "x_valid_mask": self.x_valid_mask,
             "print_valid": self.print_valid,
+            "metrics_valid": self.metrics_valid
         }
